requirements.txt
colorama==0.4.6
openai==1.14.3
python-dotenv==1.0.0
